+++ CHANGELOG.md
@@ -1,2 +1,8 @@
+- 0.1.3.0
+  - Add `Min`, `Max`, `First`, `Last`, `Option`, and `NonEmpty` instances (from `semigroups`)
+
+- 0.1.2.0
+  - Support `scientific >= 0.3.4`
+
 - 0.1.1.0
-  - Add Dual, All, Any, Sum, Product, First and Last instances
+  - Add `Dual`, `All`, `Any`, `Sum`, `Product`, `First` and `Last` instances